Set up classes with autoload for performance
