+++ bridgetown-website/Gemfile
@@ -4,6 +4,7 @@ source "https://rubygems.org"
 gem 'liquid', "> 4.0.3", github: "jaredcwhite/liquid"
 
 gem 'bridgetown-core', path: '../bridgetown-core'
+gem 'bridgetown-builder', path: '../bridgetown-builder'
 gem 'bridgetown-paginate', path: '../bridgetown-paginate'
 gem 'bridgetown', path: '../bridgetown'
